% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deviance_residuals.R
\name{plot_deviance_residuals}
\alias{plot_deviance_residuals}
\title{Function to plot the deviance residuals of a fitted coxph object.}
\usage{
plot_deviance_residuals(coxph_fit = NULL)
}
\arguments{
\item{coxph_fit}{The output from a fitted "coxph" call.}
}
\value{
A ggplot2 object presenting the deviance residuals for the model.
}
\description{
This function assists the user in assessing model performance according to deviance residuals.
The two plots produced are designed to help the analyst understand which observations are poorly predicted and how model performance fairs with respect to timing of events.
}
\examples{
library(survival)
data("heart")
# Coerce from factor
heart$transplant <- as.numeric(heart$transplant)
# Rescale age
heart$age <- heart$age+48
coxph_fit <- coxph(Surv(start, stop, event) ~
              age + transplant +surgery,
             data = heart,
             x = TRUE)

plot_deviance_residuals(coxph_fit = coxph_fit)

}
\keyword{deviance}
\keyword{fit}
\keyword{performance}
\keyword{plot}
\keyword{residuals}
\keyword{summary}
