% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dfbetas.R
\name{plot_dfbetas}
\alias{plot_dfbetas}
\title{Function to plot the DFBetas of a fitted coxph object.}
\usage{
plot_dfbetas(coxph_fit = NULL, var_names = NULL)
}
\arguments{
\item{coxph_fit}{The output from a fitted "coxph" call.}

\item{var_names}{A custom vector of variable names that correspond to the variables in the model.  Should be ordered as the covariates enter the model on the RHS of the formula.}
}
\value{
A ggplot2 object presenting the DFBetas for the model.
}
\description{
This function assists the user in assessing model performance according to DEFBetas.
# The two plots produced are designed to help the analyst understand which observations are poorly preducted and how model performance fairs with respect to timing of events.
}
\examples{
library(survival)
data("heart")
# Coerce from factor
heart$transplant <- as.numeric(heart$transplant)
# Rescale age
heart$age <- heart$age+48
fit <- coxph(Surv(start, stop, event) ~
              age + transplant +surgery,
             data = heart,
             x = TRUE)

plot_dfbetas(coxph_fit = fit, var_names = c("Age", "Transplant", "Surgery"))
}
\keyword{DFBetas}
\keyword{fit}
\keyword{performance}
\keyword{plot}
\keyword{summary}
