% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_martingale_residuals.R
\name{plot_martingale_residuals}
\alias{plot_martingale_residuals}
\title{Function to plot the deviance residuals of a fitted coxph object.}
\usage{
plot_martingale_residuals(coxph_fit = NULL, time1 = NULL,
  time2 = NULL, event = NULL)
}
\arguments{
\item{coxph_fit}{The output from a fitted "coxph" call.}

\item{time1}{This is the start time column vector used in constructing the "Surv" object.}

\item{time2}{This is the end time column vector used in constructing the "Surv" object.}

\item{event}{This is the event column vector used in constructing the "Surv" object.}
}
\value{
A ggplot2 object presenting the deviance residuals for the model.
}
\description{
This function assists the user in assessing model performance according to deviance residuals.
# The two plots produced are designed to help the analyst understand which observations are poorly preducted and how model performance fairs with respect to timing of events.
}
\examples{
library(survival)
data("heart")
# Coerce from factor
heart$transplant <- as.numeric(heart$transplant)
# Rescale age
heart$age <- heart$age+48
fit <- coxph(Surv(start, stop, event) ~
              age + transplant +surgery,
             data = heart,
             x = TRUE)

plot_martingale_residuals(coxph_fit = fit,
                          time1 = heart$start,
                          time2 = heart$stop,
                          event = heart$event)

}
\keyword{deviance}
\keyword{fit}
\keyword{performance}
\keyword{plot}
\keyword{residuals}
\keyword{summary}
