% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_martingale_residuals.R
\name{plot_martingale_residuals}
\alias{plot_martingale_residuals}
\title{Function to plot the Martingale residuals of a fitted coxph object.}
\usage{
plot_martingale_residuals(coxph_fit = NULL)
}
\arguments{
\item{coxph_fit}{The output from a fitted "coxph" call.}
}
\value{
A ggplot2 object presenting the martingale residuals for the model.
}
\description{
This function assists the user in assessing model performance according to Martingale residuals.
A list of plots.  There will be two plots for each covariate.  One showing the Martingale residuals per model with and without the covariate.
}
\examples{
library(survival)
data("heart")
# Coerce from factor
heart$transplant <- as.numeric(heart$transplant)
# Rescale age
heart$age <- heart$age+48
coxph_fit <- coxph(Surv(start, stop, event) ~
              age + transplant +surgery,
             data = heart,
             x = TRUE)

plot_list <- plot_martingale_residuals(coxph_fit = coxph_fit)

Rmisc::multiplot(plotlist = plot_list, cols = 3)

}
\keyword{fit}
\keyword{martingale}
\keyword{performance}
\keyword{plot}
\keyword{residuals}
\keyword{summary}
